# [CPP Graph] ChatGLM2 MHA support