## Type of Change: feature
API not changed

## Description
Add MHA fusion support for ChatGLM2 and avoiding initialize kv-cache twice (even with MHA fusion disabled)

## Expected Behavior & Potential Risk
N/A

## How has this PR been tested?
<details>
<summary>Test commands:</summary>

```bash
set -ex
#main branch
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=OFF -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-auto -p "$GIRL_PROMPT"
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=OFF -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-auto -p "$ZH_PROMPT1"
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=OFF -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-auto -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"<a>
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=ON -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && env ENGINE_PROFILING=1 numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 3 --memory-auto -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"<a>

#this branch
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=OFF -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-f16 -p "$GIRL_PROMPT"
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=OFF -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-auto -p "$GIRL_PROMPT"
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=OFF -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-f16 -p "$ZH_PROMPT1"
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=OFF -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-auto -p "$ZH_PROMPT1"
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=OFF -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-f16 -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"<a>
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=OFF -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-auto -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"<a>
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=ON -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && env ENGINE_PROFILING=1 numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 48 -b 2048 -c 2048 -n 3 --memory-auto -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"<a>
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=ON -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && env ENGINE_PROFILING=1 numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 3 --memory-auto -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"<a>
rm -rf bin && cmake .. -GNinja -DNE_BUILD_TESTS=ON -DNE_PROFILING=ON -DCMAKE_BUILD_TYPE=Release && ninja run_chatglm2 && env ENGINE_PROFILING=1 numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 3 --memory-f16 -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"<a>
```
</details>

In conclusion, the model can generate reasonable text with MHA fusion enabled, and it preserves the identical behavior when disable the fusion (--memory-f16). The fusion brings ~310% (4721.67ms => 1151.27ms) for first token and ~13% (47.06ms => 41.62ms) performance for next token inference.

### Main branch
- <details>
  <summary>
  <code>numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-auto -p "$GIRL_PROMPT"</code>
  </summary>

  ```
  This little girl sounds like a very adventurous and curious young [end of text]
  ```
  </details>
- <details>
  <summary>
  <code>numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-auto -p "$ZH_PROMPT1"</code>
  </summary>

  ```
  这句话不可以划分为任何一种分类。 [end of text]
  ```
  </details>
- <details>
  <summary>
  <code>numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-auto -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"</a></code>
  </summary>

  ```
  问题：河北燕郊多个小区90天卖不出一套房 置业顾问：房价“跌无可跌”

  分析：文章讲述了燕郊楼市的现状和趋势。虽然目前北京的楼市出现转热苗头，但燕郊的楼市仍然以平稳为主，房屋供应充足，可选房源多。国庆期间看房的人没有明显增加，但优质新房项目的成交比上去了。而近期社交媒体上出现了卖房热潮，有多位业主在出售燕郊的房源。

  答案：B [end of text]
  ```
  </details>
- <details>
  <summary>
  <code>ENGINE_PROFILING=1 numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 3 --memory-auto -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"</a></code>
  </summary>

  ```
  === GRAPH Profiling ===
  perf_total_per_op_us[                     ADD] =  57.482 ms
  perf_total_per_op_us[                     MUL] =  74.482 ms
  perf_total_per_op_us[                  REPEAT] = 243.032 ms
  perf_total_per_op_us[                    SILU] =  49.191 ms
  perf_total_per_op_us[                RMS_NORM] =  31.687 ms
  perf_total_per_op_us[                 MUL_MAT] = 3214.093 ms
  perf_total_per_op_us[                   SCALE] = 528.567 ms
  perf_total_per_op_us[                     CPY] =  67.748 ms
  perf_total_per_op_us[                    CONT] = 280.731 ms
  perf_total_per_op_us[                 RESHAPE] =   0.478 ms
  perf_total_per_op_us[                    VIEW] =   0.911 ms
  perf_total_per_op_us[                 PERMUTE] =   0.372 ms
  perf_total_per_op_us[                GET_ROWS] =   6.644 ms
  perf_total_per_op_us[           DIAG_MASK_INF] =  24.075 ms
  perf_total_per_op_us[                SOFT_MAX] =  88.010 ms
  perf_total_per_op_us[                    ROPE] =  47.142 ms
  perf_total_per_op_us[           INNER PRODUCT] =   1.035 ms
  ========================================
  === GRAPH Profiling ===
  perf_total_per_op_us[                     ADD] =   0.768 ms
  perf_total_per_op_us[                     MUL] =   0.937 ms
  perf_total_per_op_us[                    SILU] =   3.164 ms
  perf_total_per_op_us[                RMS_NORM] =   1.155 ms
  perf_total_per_op_us[                 MUL_MAT] =   6.039 ms
  perf_total_per_op_us[                   SCALE] =   0.717 ms
  perf_total_per_op_us[                     CPY] =   2.528 ms
  perf_total_per_op_us[                    CONT] =   0.803 ms
  perf_total_per_op_us[                 RESHAPE] =   0.272 ms
  perf_total_per_op_us[                    VIEW] =   0.862 ms
  perf_total_per_op_us[                 PERMUTE] =   0.379 ms
  perf_total_per_op_us[                GET_ROWS] =   0.029 ms
  perf_total_per_op_us[                SOFT_MAX] =   1.398 ms
  perf_total_per_op_us[                    ROPE] =   1.399 ms
  perf_total_per_op_us[           INNER PRODUCT] =  24.177 ms
  ========================================
  === GRAPH Profiling ===
  perf_total_per_op_us[                     ADD] =   0.679 ms
  perf_total_per_op_us[                     MUL] =   0.941 ms
  perf_total_per_op_us[                    SILU] =   3.047 ms
  perf_total_per_op_us[                RMS_NORM] =   1.097 ms
  perf_total_per_op_us[                 MUL_MAT] =   5.804 ms
  perf_total_per_op_us[                   SCALE] =   0.682 ms
  perf_total_per_op_us[                     CPY] =   2.397 ms
  perf_total_per_op_us[                    CONT] =   0.494 ms
  perf_total_per_op_us[                 RESHAPE] =   0.268 ms
  perf_total_per_op_us[                    VIEW] =   0.831 ms
  perf_total_per_op_us[                 PERMUTE] =   0.373 ms
  perf_total_per_op_us[                GET_ROWS] =   0.029 ms
  perf_total_per_op_us[                SOFT_MAX] =   1.319 ms
  perf_total_per_op_us[                    ROPE] =   1.382 ms
  perf_total_per_op_us[           INNER PRODUCT] =  23.378 ms
  ========================================
  model_print_timings:        load time =  5957.46 ms
  model_print_timings:      sample time =     3.49 ms /     3 runs   (    1.16 ms per token)
  model_print_timings: prompt eval time =  4721.67 ms /  1323 tokens (    3.57 ms per token)
  model_print_timings:        eval time =    96.11 ms /     2 runs   (   48.05 ms per token)
  model_print_timings:       total time =  6108.33 ms
  ========== eval time log of each prediction ==========
  prediction   0, time: 4721.67ms
  prediction   1, time: 49.04ms
  prediction   2, time: 47.06ms
  ```
  </details>

### This branch
- <details>
  <summary>
  <code>numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-f16 -p "$GIRL_PROMPT"</code>
  </summary>

  ```
  This little girl sounds like a very adventurous and curious young [end of text]
  ```
  </details>
- <details>
  <summary>
  <code>numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-auto -p "$GIRL_PROMPT"</code>
  </summary>

  ```
  This little girl sounds like a very adventurous and curious young [end of text]
  ```
  </details>
- <details>
  <summary>
  <code>numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-f16 -p "$ZH_PROMPT1"</code>
  </summary>

  ```
  这句话不可以划分为任何一种分类。 [end of text]
  ```
  </details>
- <details>
  <summary>
  <code>numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-auto -p "$ZH_PROMPT1"</code>
  </summary>

  ```
  这句话不可以划分为任何一种分类。 [end of text]
  ```
  </details>
- <details>
  <summary>
  <code>numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-f16 -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"</a></code>
  </summary>

  ```
  问题：河北燕郊多个小区90天卖不出一套房 置业顾问：房价“跌无可跌”

  分析：文章讲述了燕郊楼市的现状和趋势。虽然目前北京的楼市出现转热苗头，但燕郊的楼市仍然以平稳为主，房屋供应充足，可选房源多。国庆期间看房的人没有明显增加，但优质新房项目的成交比上去了。而近期社交媒体上出现了卖房热潮，有多位业主在出售燕郊的房源。

  答案：B [end of text]
  ```
  </details>
- <details>
  <summary>
  <code>numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 256 --memory-auto -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"</a></code>
  </summary>

  ```
  近几个月来,河北燕郊多个小区房价下跌严重,许多房屋已经卖了超过90天,但仍然没有找到买家。置业顾问表示,房价已经“跌无可跌”,现在购房者的需求主要是自住需求。不过,房地产市场前景依然不乐观,去库存迫在眉睫,房价下跌趋势仍将继续。 [end of text]
  ```
  </details>
- <details>
  <summary>
  <code>ENGINE_PROFILING=1 numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 48 -b 2048 -c 2048 -n 3 --memory-auto -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"</a></code>
  </summary>

  ```
  === GRAPH Profiling ===
  perf_total_per_op_us[                     ADD] =  44.110 ms
  perf_total_per_op_us[                     MUL] =  71.412 ms
  perf_total_per_op_us[                  REPEAT] = 236.015 ms
  perf_total_per_op_us[                    SILU] =  55.165 ms
  perf_total_per_op_us[                RMS_NORM] =  29.132 ms
  perf_total_per_op_us[                 MUL_MAT] = 632.301 ms
  perf_total_per_op_us[                    VIEW] =   1.173 ms
  perf_total_per_op_us[                 PERMUTE] =   0.095 ms
  perf_total_per_op_us[                GET_ROWS] =   6.754 ms
  perf_total_per_op_us[                    ROPE] =  46.325 ms
  perf_total_per_op_us[              FLASH_ATTN] =  82.043 ms
  perf_total_per_op_us[    FLASH_ATTN_KV_UPDATE] =  13.944 ms
  perf_total_per_op_us[           INNER PRODUCT] =   1.045 ms
  ========================================
  === GRAPH Profiling ===
  perf_total_per_op_us[                     ADD] =   0.717 ms
  perf_total_per_op_us[                     MUL] =   0.956 ms
  perf_total_per_op_us[                    SILU] =   3.043 ms
  perf_total_per_op_us[                RMS_NORM] =   1.211 ms
  perf_total_per_op_us[                    VIEW] =   1.324 ms
  perf_total_per_op_us[                 PERMUTE] =   0.089 ms
  perf_total_per_op_us[                GET_ROWS] =   0.030 ms
  perf_total_per_op_us[                    ROPE] =   1.298 ms
  perf_total_per_op_us[              FLASH_ATTN] =   3.667 ms
  perf_total_per_op_us[    FLASH_ATTN_KV_UPDATE] =   2.044 ms
  perf_total_per_op_us[           INNER PRODUCT] =  23.879 ms
  ========================================
  === GRAPH Profiling ===
  perf_total_per_op_us[                     ADD] =   0.636 ms
  perf_total_per_op_us[                     MUL] =   0.881 ms
  perf_total_per_op_us[                    SILU] =   2.886 ms
  perf_total_per_op_us[                RMS_NORM] =   1.139 ms
  perf_total_per_op_us[                    VIEW] =   1.523 ms
  perf_total_per_op_us[                 PERMUTE] =   0.088 ms
  perf_total_per_op_us[                GET_ROWS] =   0.017 ms
  perf_total_per_op_us[                    ROPE] =   1.324 ms
  perf_total_per_op_us[              FLASH_ATTN] =   3.421 ms
  perf_total_per_op_us[    FLASH_ATTN_KV_UPDATE] =   1.969 ms
  perf_total_per_op_us[           INNER PRODUCT] =  23.324 ms
  ========================================
  
  model_print_timings:        load time =  2456.21 ms
  model_print_timings:      sample time =     3.36 ms /     3 runs   (    1.12 ms per token)
  model_print_timings: prompt eval time =  1223.68 ms /  1323 tokens (    0.92 ms per token)
  model_print_timings:        eval time =    81.88 ms /     2 runs   (   40.94 ms per token)
  model_print_timings:       total time =  2593.04 ms
  ========== eval time log of each prediction ==========
  prediction   0, time: 1223.68ms
  prediction   1, time: 41.54ms
  prediction   2, time: 40.34ms
  ```
  </details>
- <details>
  <summary>
  <code>ENGINE_PROFILING=1 numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 3 --memory-auto -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"</a></code>
  </summary>

  ```
  === GRAPH Profiling ===
  perf_total_per_op_us[                     ADD] =  45.072 ms
  perf_total_per_op_us[                     MUL] =  69.500 ms
  perf_total_per_op_us[                  REPEAT] = 240.945 ms
  perf_total_per_op_us[                    SILU] =  52.971 ms
  perf_total_per_op_us[                RMS_NORM] =  27.820 ms
  perf_total_per_op_us[                 MUL_MAT] = 562.168 ms
  perf_total_per_op_us[                    VIEW] =   1.297 ms
  perf_total_per_op_us[                 PERMUTE] =   0.094 ms
  perf_total_per_op_us[                GET_ROWS] =   5.984 ms
  perf_total_per_op_us[                    ROPE] =  46.332 ms
  perf_total_per_op_us[              FLASH_ATTN] =  78.879 ms
  perf_total_per_op_us[    FLASH_ATTN_KV_UPDATE] =  14.803 ms
  perf_total_per_op_us[           INNER PRODUCT] =   1.036 ms
  ========================================
  === GRAPH Profiling ===
  perf_total_per_op_us[                     ADD] =   0.771 ms
  perf_total_per_op_us[                     MUL] =   0.976 ms
  perf_total_per_op_us[                    SILU] =   3.174 ms
  perf_total_per_op_us[                RMS_NORM] =   1.212 ms
  perf_total_per_op_us[                    VIEW] =   1.384 ms
  perf_total_per_op_us[                 PERMUTE] =   0.093 ms
  perf_total_per_op_us[                GET_ROWS] =   0.032 ms
  perf_total_per_op_us[                    ROPE] =   1.463 ms
  perf_total_per_op_us[              FLASH_ATTN] =   3.748 ms
  perf_total_per_op_us[    FLASH_ATTN_KV_UPDATE] =   2.281 ms
  perf_total_per_op_us[           INNER PRODUCT] =  23.935 ms
  ========================================
  === GRAPH Profiling ===
  perf_total_per_op_us[                     ADD] =   0.765 ms
  perf_total_per_op_us[                     MUL] =   0.931 ms
  perf_total_per_op_us[                    SILU] =   3.073 ms
  perf_total_per_op_us[                RMS_NORM] =   1.127 ms
  perf_total_per_op_us[                    VIEW] =   1.617 ms
  perf_total_per_op_us[                 PERMUTE] =   0.094 ms
  perf_total_per_op_us[                GET_ROWS] =   0.030 ms
  perf_total_per_op_us[                    ROPE] =   1.427 ms
  perf_total_per_op_us[              FLASH_ATTN] =   3.689 ms
  perf_total_per_op_us[    FLASH_ATTN_KV_UPDATE] =   2.255 ms
  perf_total_per_op_us[           INNER PRODUCT] =  23.472 ms
  ========================================
  
  model_print_timings:        load time =  2333.29 ms
  model_print_timings:      sample time =     3.37 ms /     3 runs   (    1.12 ms per token)
  model_print_timings: prompt eval time =  1151.27 ms /  1323 tokens (    0.87 ms per token)
  model_print_timings:        eval time =    84.09 ms /     2 runs   (   42.05 ms per token)
  model_print_timings:       total time =  2472.01 ms
  ========== eval time log of each prediction ==========
  prediction   0, time: 1151.27ms
  prediction   1, time: 42.47ms
  prediction   2, time: 41.62ms
  ```
  </details>
- <details>
  <summary>
  <code>ENGINE_PROFILING=1 numactl -m 1 -C 56-111 bin/run_chatglm2 -m /home_local/dingyi/data/chatglm2-6b-pr410-q4j-sym-int8-fp32-g128.bin --seed 1234 -t 56 -b 2048 -c 2048 -n 3 --memory-f16 -p <a href="https://pastebin.com/v5pXFdCg">"$ZH_PROMPT2"</a></code>
  </summary>

  ```
  perf_total_per_op_us[                     ADD] =  47.564 ms
  perf_total_per_op_us[                     MUL] =  79.097 ms
  perf_total_per_op_us[                  REPEAT] = 238.934 ms
  perf_total_per_op_us[                    SILU] =  49.355 ms
  perf_total_per_op_us[                RMS_NORM] =  36.452 ms
  perf_total_per_op_us[                 MUL_MAT] = 3119.894 ms
  perf_total_per_op_us[                   SCALE] = 531.817 ms
  perf_total_per_op_us[                     CPY] =   5.305 ms
…