@DDEle 2023-10-11
> BTW I think its better to summary the results as an table in PR description and attach detailed log in the following comments

Make sense to me. I'll put generated text together next time.